Выполняет модуляцию
