+++ js/constants.js
@@ -1,0 +1,64 @@
+const PHOTOS = 25;
+
+const DESCRIPTION = [
+  'Классно!',
+  'Бодро',
+  'Огонь!',
+  'Пожар',
+  'Весело',
+  'Вкусно',
+  'Аппетитно',
+  'Восторженно',
+  'Смело',
+  'Отнудь не...'
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const MIN_LIKES = 15;
+
+const MAX_LIKES = 200;
+
+const COMMENTS = 500;
+
+const MIN_AVATAR = 1;
+
+const MAX_AVATAR = 6;
+
+const NAMES = [
+  'Иван',
+  'Пётр',
+  'Александр',
+  'Николай',
+  'Максим',
+  'Светлана',
+  'Елена',
+  'Мария',
+  'Евгения',
+  'Надежда'
+];
+
+const MIN_COMMENTS = 1;
+
+const MAX_COMMENTS = 30;
+
+export {
+  PHOTOS,
+  DESCRIPTION,
+  MESSAGE,
+  MIN_LIKES,
+  MAX_LIKES,
+  COMMENTS,
+  MIN_AVATAR,
+  MAX_AVATAR,
+  NAMES,
+  MIN_COMMENTS,
+  MAX_COMMENTS
+};
